feat: moved validateQuoteSchema as public static to Quote
