+++ src/ctype/CTypeUtils.ts
@@ -9,7 +9,6 @@ import Ajv from 'ajv'
 import { CTypeModel } from './CTypeSchema'
 import ICType from '../types/CType'
 import Crypto from '../crypto'
-import QuoteSchema from '../quote/QuoteSchema'
 
 export function verifySchemaWithErrors(
   model: any,
@@ -42,22 +41,4 @@ export function verifyClaimStructure(claim: any, schema: any): boolean {
 
 export function getHashForSchema(schema: ICType['schema']): string {
   return Crypto.hashObjectAsStr(schema)
-}
-
-export function validateQuoteSchema(
-  schema: object,
-  validate: object,
-  messages?: string[]
-): boolean | PromiseLike<any> {
-  const ajv = new Ajv()
-  ajv.addMetaSchema(QuoteSchema)
-  const result = ajv.validate(schema, validate)
-  if (!result && ajv.errors) {
-    if (messages) {
-      ajv.errors.forEach((error: any) => {
-        messages.push(error.message)
-      })
-    }
-  }
-  return !!result
 }
